Now addons are downloaded from the vova1234's site
May work / work incompletely / not work

+++ WorkshopDownloader.Core/Zip/Unzipper.cs
@@ -1,21 +1,34 @@
 ﻿using System.IO;
 using System.IO.Compression;
 using System.Threading.Tasks;
+using System.Windows.Forms;
 
 namespace WorkshopDownloader.Core.ZipUtils
 {
     public static class Unzipper
     {
-        public static async void UnzipFileAsync(string filePath)
+        public static async void UnzipFileAsync(string filePath, bool createFolder = true)
         {
             string extractPath = filePath.Replace(".zip", "");
 
-            if (Directory.Exists(extractPath))
-                Directory.Delete(extractPath, true);
+            if (createFolder)
+            {
+                Directory.CreateDirectory(extractPath);
+                if (Directory.Exists(extractPath)) Directory.Delete(extractPath, true);
+            }
+            else
+            {
+                string[] t = extractPath.Split('\\');
+                extractPath = string.Empty;
+                for (int i = 0; i < t.Length; i++)
+                {
+                    if (i == t.Length - 1) break;
+                    extractPath += t[i] + "\\";
+                }
+            }
 
-            Directory.CreateDirectory(extractPath);
             await Task.Run(() => ZipFile.ExtractToDirectory(filePath, extractPath));
-
+            
             File.Delete(filePath);
         }
     }
